I updated the content and tried to fix things.
I attempted to fix the bug where the line under the title wasn't showing, I moved the comment and changed it because it was showing on the page, I added the font sans-serif, I also attempted to curve the corners of the image, and I added another image. I also changed the alt of the first image to make Charles seem less dangerous but more annoying. Plus, I added a favicon and I don't know if it'll work.